refactor: 게임 로직 Map에서 Redis로 리팩토링

+++ backend/module-core/src/main/java/dev/gunlog/multi/dto/RoomCreateResponseDto.java
@@ -7,10 +7,10 @@ import lombok.Setter;
 @Getter
 @Setter
 public class RoomCreateResponseDto {
-    private Integer roomId;
+    private Long roomId;
 
     @Builder
-    public RoomCreateResponseDto(Integer roomId) {
+    public RoomCreateResponseDto(Long roomId) {
         this.roomId = roomId;
     }
 }
